refactor(queues): move negotiation-timeout into negotiations/ subdir
